requirements.txt
absl-py==0.10.0
astroid==2.4.2
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.33.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.3.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tensorflow-cpu==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
